+++ src/read.cpp
@@ -30,7 +30,7 @@ string Read::toString() {
 }
 
 
-void Read::addVariant(int position, int allele, vector<float> em, int quality) {
+void Read::addVariant(int position, int allele, vector<unsigned int> em, int quality) {
 	variants.push_back(enriched_entry_t(position, allele, em, quality));
 }
 
@@ -110,7 +110,7 @@ std::vector<float> Read::getEmissionProbability(size_t variant_idx) const {
 }
 
 
-void Read::setEmissionProbability(size_t variant_idx, std::vector<float> emission) {
+void Read::setEmissionProbability(size_t variant_idx, std::vector<unsigned int> emission) {
 	assert(variant_idx < variants.size());
 	variants[variant_idx].entry.set_emission_score(emission);
 }
