+++ src/read.h
@@ -12,7 +12,7 @@ public:
 	Read(const std::string& name, int mapq, int source_id, int sample_id, int reference_start = -1, const std::string& BX_tag = "");
 	virtual ~Read() {}
 	std::string toString();
-	void addVariant(int position, int allele, std::vector<float> em, int quality);
+	void addVariant(int position, int allele, std::vector<unsigned int> em, int quality);
 	void sortVariants();
 	/** Returns the position of the first variant. **/
 	int firstPosition() const;
@@ -27,7 +27,7 @@ public:
 	int getAllele(size_t variant_idx) const;
 	void setAllele(size_t variant_idx, int allele);
 	std::vector<float> getEmissionProbability(size_t variant_idx) const;
-	void setEmissionProbability(size_t variant_idx, std::vector<float> emission);
+	void setEmissionProbability(size_t variant_idx, std::vector<unsigned int> emission);
 	int getQuality(size_t variant_idx) const;
 	void setQuality(size_t variant_idx, int quality);
 	const Entry* getEntry(size_t variant_idx) const;
@@ -45,7 +45,7 @@ private:
 	typedef struct enriched_entry_t {
 		Entry entry;
 		int position;
-		enriched_entry_t(int position, int allele, std::vector<float> em, int quality) :
+		enriched_entry_t(int position, int allele, std::vector<unsigned int> em, int quality) :
 			entry(0,allele,em,quality), position(position) {}
 	} enriched_entry_t;
